LSP user keybindings to match VS Code as much as possible.
Bound most LSP commands to match VS Code's for easy transfer where
possible without overriding any ST keybindings. I checked log outputs
of commands I was going to bind to a LSP command to see if ST bound it
first, and if it did would try an alternate, e.g., use Alt combination
instead.

In one case, F2, I overrode the ST binding which is for bookmarks
because I don't use them currently, and because alternates like Ctrl-F2
and Alt-F2 are also bound by ST to bookmark-related things, and F2 is
pretty common among all editors/IDEs for renaming, so it made sense to
ditch ST binding to it in order to use it for LSP rename capability.

I turned off ST completions when LSP server present to reduce candidates
, but this may have an effect on ST snippets, it remains to be seen, so
this may have to revert.

SublimeLinter plugins for pylint and standard are gone. I'm moving much
of that stuff towards LSP as possible.

Indent guides are on because in js/jsx functions I sometimes get lost
finding which if-else closes what etc.

"scroll_context_lines" option was fixed in 4085 so I turned it back on.